+++ src/index.ts
@@ -1,3 +1,12 @@
 export * from "./components/voiceProvider.js";
 export * from "./components/useVoice.js";
 export * from "./components/useVoiceControls.js";
+export {
+  VoiceProvider,
+  useVoiceContext,
+  WithVoice,
+} from "./components/voiceProvider";
+export { useVoice } from "./components/useVoice";
+export { useVoiceControls } from "./components/useVoiceControls";
+export { VoiceIntents, useAutoVoiceIntents } from "./components/voiceIntents";
+export type { IntentSpec, VoiceIntentsProps } from "./components/voiceIntents";