+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on **[Keep a Changelog](https://keepachangelog.com/en/1.1.0/
 ## [Unreleased]
 
 - **Added**
-  - (placeholder)
+  - VoiceIntents auto-register/unregister mechanism added
 
 - **Changed**
   - README.md update